Fix concurrent_computing

+++ src/concurrent_computing.rs
@@ -10,7 +10,7 @@ fn main() {
     for s in strings.move_iter(){
         spawn(proc() {
             // We use a random u8 (so an integer from 0 to 255)
-            sleep(Duration::milliseconds(random::<u8>() as i32));
+            sleep(Duration::milliseconds(random::<u8>() as i64));
             println!("{}", s);
         });
     }
